requirements.txt
fastapi
uvicorn[standard]

pydantic
python-dotenv

fal-client
replicate

httpx
openai